% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{acidData-package}
\alias{acidData}
\alias{acidData-package}
\title{acidData
Reference data from Acid Genomics.}
\description{
Reference data from Acid Genomics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/acidgenomics/acidData/}
  \item Report bugs at \url{https://github.com/acidgenomics/acidData/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (0000-0002-2403-2221)

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
