% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{acidtest-package}
\alias{acidtest}
\alias{acidtest-package}
\title{acidtest
Reference data for unit testing from Acid Genomics.}
\description{
Reference data for unit testing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://acidtest.acidgenomics.com/}
  \item \url{https://github.com/acidgenomics/acidtest/}
  \item Report bugs at \url{https://github.com/acidgenomics/acidtest/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (0000-0002-2403-2221)

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
