% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\docType{data}
\name{band_instruments}
\alias{band_instruments}
\title{Band membership}
\format{An object of class \code{DFrame} with 3 rows and 2 columns.}
\usage{
data(band_instruments)
}
\value{
\code{DFrame}.
}
\description{
These data sets describe band members of the Beatles and Rolling Stones. They
are toy data sets that can be displayed in their entirety on a slide (e.g. to
demonstrate a join).
}
\note{
Updated 2020-01-28.
}
\examples{
if (requireNamespace("S4Vectors", quietly = TRUE)) {
    data(band_instruments)
    class(band_instruments)
    band_instruments
}
}
\seealso{
\code{help(topic = "band_instruments", package = "dplyr")}
}
\keyword{datasets}
