% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\docType{data}
\name{correlation}
\alias{correlation}
\title{Correlation}
\format{An object of class \code{list} of length 6.}
\usage{
data(correlation)
}
\value{
\code{list}.
}
\description{
Example \code{list} data used to test \code{correlation()} methods.
}
\note{
Updated 2020-01-28.
}
\examples{
data(correlation)
class(correlation)
names(correlation)
}
\keyword{datasets}
