% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\docType{data}
\name{data.frame}
\alias{data.frame}
\title{Data frame}
\format{An object of class \code{data.frame} with 4 rows and 4 columns.}
\usage{
data.frame
}
\description{
Example \code{data.frame} object.
}
\examples{
data(data.frame)
class(data.frame)
data.frame
}
\keyword{datasets}
