% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.table.R
\docType{data}
\name{data.table}
\alias{data.table}
\title{Data table}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 4 rows and 5 columns.
}
\usage{
data(data.table)
}
\value{
\code{data.table}.
}
\description{
Example \code{data.table} object.
}
\note{
Updated 2020-01-28.
}
\examples{
if (requireNamespace("data.table", quietly = TRUE)) {
    data(data.table)
    class(data.table)
    data.table
}
}
\keyword{datasets}
