% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrame.R
\docType{data}
\name{DataFrame}
\alias{DataFrame}
\title{S4 data frame}
\format{
An object of class \code{DFrame} with 4 rows and 2 columns.
}
\usage{
data(DataFrame)
}
\value{
\code{DFrame} / \code{DataFrame}.
}
\description{
Example Bioconductor S4 \code{DataFrame} object.
}
\note{
Updated 2021-10-14.
}
\examples{
if (requireNamespace("S4Vectors", quietly = TRUE)) {
    data(DataFrame)
    class(DataFrame)
    is(DataFrame, "DataFrame")
    DataFrame
}
}
\keyword{datasets}
