% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df}
\alias{df}
\title{Data frame}
\format{An object of class \code{DataFrame} with 4 rows and 4 columns.}
\usage{
df
}
\description{
Example Bioconductor S4 \code{DataFrame} object.
}
\examples{
if (requireNamespace("S4Vectors", quietly = TRUE)) {
    data(df)
    class(df)
    df
}
}
\keyword{datasets}
