% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFrame.R
\docType{data}
\name{DFrame}
\alias{DFrame}
\title{S4 data frame}
\format{
An object of class \code{DFrame} with 4 rows and 2 columns.
}
\usage{
data(DFrame)
}
\value{
\code{DFrame}.
}
\description{
Example Bioconductor S4 \code{DFrame} object.
}
\note{
Updated 2023-04-26.
}
\examples{
if (requireNamespace("S4Vectors", quietly = TRUE)) {
    data(DFrame)
    class(DFrame)
    is(DFrame, "DFrame")
    DFrame
}
}
\keyword{datasets}
