% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFrameList.R
\docType{data}
\name{DFrameList}
\alias{DFrameList}
\title{S4 data frame list}
\format{
An object of class \code{SimpleDFrameList} of dimension 1 x 2 x 2 x 2.
}
\usage{
data(DFrameList)
}
\value{
\code{DFrameList}.
}
\description{
Example Bioconductor S4 \code{DFrameList} object.
}
\note{
Updated 2023-11-08.
}
\examples{
if (requireNamespace("IRanges", quietly = TRUE)) {
    data(DFrameList)
    class(DFrameList)
    is(DFrameList, "DFrameList")
    DFrameList
}
}
\keyword{datasets}
