% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt}
\alias{dt}
\title{Data table}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 4 rows and 5 columns.}
\usage{
dt
}
\description{
Example \code{data.table} object.
}
\examples{
if (requireNamespace("data.table", quietly = TRUE)) {
    data(dt)
    class(dt)
    dt
}
}
\keyword{datasets}
