% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicRanges.R
\docType{data}
\name{GenomicRanges}
\alias{GenomicRanges}
\title{Genomic ranges}
\format{
An object of class \code{GRanges} of length 5.
}
\usage{
data(GenomicRanges)
}
\value{
\code{GRanges} / \code{GenomicRanges}.
}
\description{
Example \code{GenomicRanges} object.
}
\note{
Updated 2021-10-14.
}
\examples{
if (requireNamespace("GenomicRanges", quietly = TRUE)) {
    data(GenomicRanges)
    class(GenomicRanges)
    is(GenomicRanges, "GenomicRanges")
    GenomicRanges
}
}
\keyword{datasets}
