% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gr}
\alias{gr}
\title{Example GRanges
Genomic ranges.}
\format{An object of class \code{GRanges} of length 5.}
\usage{
gr
}
\description{
Example GRanges
Genomic ranges.
}
\examples{
data(gr)
gr
}
\keyword{datasets}
