% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges.R
\docType{data}
\name{GRanges}
\alias{GRanges}
\title{Genomic ranges}
\format{
An object of class \code{GRanges} of length 5.
}
\usage{
data(GRanges)
}
\value{
\code{GRanges}.
}
\description{
Example \code{GRanges} object.
}
\note{
Updated 2023-04-26.
}
\examples{
if (requireNamespace("GenomicRanges", quietly = TRUE)) {
    data(GRanges)
    class(GRanges)
    is(GRanges, "GRanges")
    print(GRanges)
}
}
\keyword{datasets}
