% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesList.R
\docType{data}
\name{GRangesList}
\alias{GRangesList}
\title{S4 genomic ranges list}
\format{
An object of class \code{CompressedGRangesList} of length 3.
}
\usage{
data(GRangesList)
}
\value{
\code{GRangesList}.
}
\description{
Example Bioconductor S4 \code{GRangesList} object.
}
\note{
Updated 2023-11-08.
}
\examples{
if (requireNamespace("GenomicRanges", quietly = TRUE)) {
    data(GRangesList)
    class(GRangesList)
    is(GRangesList, "GRangesList")
    print(GRangesList)
}
}
\keyword{datasets}
