% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegerRanges.R
\docType{data}
\name{IntegerRanges}
\alias{IntegerRanges}
\title{Integer ranges}
\format{
An object of class \code{IRanges} of length 3.
}
\usage{
data(IntegerRanges)
}
\value{
\code{IRanges} / \code{IntegerRanges}.
}
\description{
Example \code{IntegerRanges} object.
}
\note{
Updated 2021-10-14.
}
\examples{
if (requireNamespace("IRanges", quietly = TRUE)) {
    data(IntegerRanges)
    class(IntegerRanges)
    is(IntegerRanges, "IntegerRanges")
    IntegerRanges
}
}
\keyword{datasets}
