% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ir}
\alias{ir}
\title{Integer ranges}
\format{An object of class \code{IRanges} of length 3.}
\usage{
ir
}
\description{
Examples \code{IRanges} object.
}
\examples{
data(ir)
class(ir)
ir
}
\keyword{datasets}
