% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRanges.R
\docType{data}
\name{IRanges}
\alias{IRanges}
\title{Integer ranges}
\format{
An object of class \code{IRanges} of length 3.
}
\usage{
data(IRanges)
}
\value{
\code{IRanges}.
}
\description{
Example \code{IRanges} object.
}
\note{
Updated 2023-04-26.
}
\examples{
if (requireNamespace("IRanges", quietly = TRUE)) {
    data(IRanges)
    class(IRanges)
    is(IRanges, "IRanges")
    print(IRanges)
}
}
\keyword{datasets}
