% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\docType{data}
\name{join}
\alias{join}
\title{Test data for join operations}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(join)
}
\value{
\code{list}.
}
\description{
Test data for join operations
}
\note{
Updated 2021-10-14.
}
\examples{
if (requireNamespace("S4Vectors", quietly = TRUE)) {
    data(join)
    class(join)
    names(join)
}
}
\keyword{datasets}
