% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnownMarkers.R
\docType{data}
\name{KnownMarkers}
\alias{KnownMarkers}
\title{Known markers example}
\format{
An object of class \code{KnownMarkers} with 10 rows and 10 columns.
}
\usage{
data(KnownMarkers)
}
\value{
\code{KnownMarkers}.
}
\description{
Known markers example
}
\note{
Updated 2022-03-04.
}
\examples{
if (requireNamespace("AcidSingleCell", quietly = TRUE)) {
    data(KnownMarkers)
    print(KnownMarkers)
}
}
\keyword{datasets}
