% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lfc}
\alias{lfc}
\title{Log fold change}
\format{An object of class \code{matrix} with 8 rows and 4 columns.}
\usage{
lfc
}
\description{
Example \code{matrix} containing log base 2 fold change values.
}
\examples{
data(lfc)
class(lfc)
lfc
}
\keyword{datasets}
