% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\docType{data}
\name{matrix}
\alias{matrix}
\title{Matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 4 columns.
}
\usage{
data(matrix)
}
\value{
\code{matrix}.
}
\description{
Example \code{matrix} object.
}
\note{
Updated 2020-01-28.
}
\examples{
data(matrix)
class(matrix)
matrix
}
\keyword{datasets}
