% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_lfc.R
\docType{data}
\name{matrix_lfc}
\alias{matrix_lfc}
\title{Log fold change}
\format{An object of class \code{matrix} with 8 rows and 4 columns.}
\usage{
matrix_lfc
}
\description{
Example \code{matrix} containing log base 2 fold change values.
}
\note{
Updated 2020-01-18.
}
\examples{
data(matrix_lfc)
class(matrix_lfc)
matrix_lfc
}
\keyword{datasets}
