% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mn}
\alias{mn}
\title{Make names}
\format{An object of class \code{list} of length 7.}
\usage{
mn
}
\description{
Example \code{list} data used to test syntactic naming functions.
}
\examples{
data(mn)
class(mn)
names(mn)
}
\keyword{datasets}
