% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce}
\alias{sce}
\title{Single cell experiment}
\format{An object of class \code{SingleCellExperiment} with 230 rows and 80 columns.}
\usage{
sce
}
\description{
Example \code{SingleCellExperiment} containing gene-level RNA-seq data.
}
\examples{
if (requireNamespace("SingleCellExperiment", quietly = TRUE)) {
    data(sce)
    class(sce)
    sce
}
}
\seealso{
\code{scater::sc_example_counts}.
}
\keyword{datasets}
