% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seurat}
\alias{seurat}
\title{Seurat}
\format{An object of class \code{Seurat} with 230 rows and 80 columns.}
\usage{
seurat
}
\description{
Example \code{Seurat} (v3) object.
}
\examples{
if (requireNamespace("Seurat", quietly = TRUE)) {
    data(seurat)
    class(seurat)
    seurat
}
}
\keyword{datasets}
