% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratMarkersPerCluster.R
\docType{data}
\name{SeuratMarkersPerCluster}
\alias{SeuratMarkersPerCluster}
\title{Seurat markers per cluster}
\format{
An object of class \code{SeuratMarkersPerCluster} of dimension 106 x 79 x 37 x 8 x 8 x 8.
}
\usage{
data(SeuratMarkersPerCluster)
}
\value{
\code{SeuratMarkersPerCluster}.
}
\description{
Seurat markers per cluster
}
\note{
Updated 2022-03-04.
}
\examples{
data(SeuratMarkersPerCluster)
print(SeuratMarkersPerCluster)
}
\keyword{datasets}
