% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellExperiment.R
\docType{data}
\name{SingleCellExperiment}
\alias{SingleCellExperiment}
\title{Single cell experiment}
\format{An object of class \code{SingleCellExperiment} with 500 rows and 100 columns.}
\usage{
SingleCellExperiment
}
\description{
Example \code{SingleCellExperiment} containing gene-level RNA-seq data.
}
\examples{
if (requireNamespace("SingleCellExperiment", quietly = TRUE)) {
    data(SingleCellExperiment)
    class(SingleCellExperiment)
    SingleCellExperiment
}
}
\seealso{
\itemize{
\item \code{splatter::splatSimulate()}.
\item \code{scater::sc_example_counts}.
}
}
\keyword{datasets}
