% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellExperiment_lanesplit.R
\docType{data}
\name{SingleCellExperiment_lanesplit}
\alias{SingleCellExperiment_lanesplit}
\title{Single cell experiment with lane-split technical replicates}
\format{
An object of class \code{SingleCellExperiment} with 100 rows and 32 columns.
}
\usage{
data(SingleCellExperiment_lanesplit)
}
\value{
\code{SingleCellExperiment}.
}
\description{
Example \code{SingleCellExperiment} containing gene-level RNA-seq data and
lane-split technical replicates.
}
\note{
Updated 2022-03-10.
}
\examples{
if (requireNamespace("SingleCellExperiment", quietly = TRUE)) {
    data(SingleCellExperiment_lanesplit)
    class(SingleCellExperiment_lanesplit)
    print(SingleCellExperiment_lanesplit)
}
}
\keyword{datasets}
