% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellExperiment_Seurat.R
\docType{data}
\name{SingleCellExperiment_Seurat}
\alias{SingleCellExperiment_Seurat}
\title{Single cell experiment}
\format{
An object of class \code{SingleCellExperiment} with 230 rows and 80 columns.
}
\usage{
data(SingleCellExperiment_Seurat)
}
\value{
\code{SingleCellExperiment}.
}
\description{
Example \code{SingleCellExperiment} containing gene-level RNA-seq data.
}
\note{
Updated 2022-03-04.
}
\examples{
if (requireNamespace("SingleCellExperiment", quietly = TRUE)) {
    data(SingleCellExperiment_Seurat)
    class(SingleCellExperiment_Seurat)
    SingleCellExperiment_Seurat
}
}
\seealso{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{`data(pbmc_small, package = "Seurat")`.
}\if{html}{\out{</div>}}
}
\keyword{datasets}
