% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseMatrix.R
\docType{data}
\name{sparseMatrix}
\alias{sparseMatrix}
\title{Sparse matrix}
\format{An object of class \code{dgCMatrix} with 8 rows and 10 columns.}
\usage{
sparseMatrix
}
\description{
Example \code{sparseMatrix} object.
}
\examples{
if (requireNamespace("Matrix", quietly = TRUE)) {
    data(sparseMatrix)
    class(sparseMatrix)
    sparseMatrix
}
}
\keyword{datasets}
