% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitDFrameList.R
\docType{data}
\name{SplitDFrameList}
\alias{SplitDFrameList}
\title{S4 split data frame list}
\format{
An object of class \code{CompressedSplitDFrameList} of dimension 1 x 2 x 2 x 2.
}
\usage{
data(SplitDFrameList)
}
\value{
\code{SplitDFrameList}.
}
\description{
Example Bioconductor S4 \code{SplitDFrameList} object.
}
\note{
Updated 2023-11-08.
}
\examples{
if (requireNamespace("IRanges", quietly = TRUE)) {
    data(SplitDFrameList)
    class(SplitDFrameList)
    is(SplitDFrameList, "SplitDFrameList")
    print(SplitDFrameList)
}
}
\keyword{datasets}
