% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperiment.R
\docType{data}
\name{SummarizedExperiment}
\alias{SummarizedExperiment}
\title{Summarized experiment}
\format{An object of class \code{SummarizedExperiment} with 50 rows and 4 columns.}
\usage{
data(SummarizedExperiment)
}
\value{
\code{SummarizedExperiment}.
}
\description{
Modified version of the \code{SummarizedExperiment()} documentation example.
}
\note{
Updated 2020-01-28.
}
\examples{
if (requireNamespace("SummarizedExperiment", quietly = TRUE)) {
    data(SummarizedExperiment)
    class(SummarizedExperiment)
    SummarizedExperiment
}
}
\keyword{datasets}
