% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_df.R
\docType{data}
\name{tbl_df}
\alias{tbl_df}
\title{Tibble}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 5 columns.}
\usage{
tbl_df
}
\description{
Example \code{tbl_df} object.
}
\note{
Updated 2020-01-18.
}
\examples{
if (requireNamespace("tibble", quietly = TRUE)) {
    data(tbl_df)
    class(tbl_df)
    tbl_df
}
}
\keyword{datasets}
