% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\docType{data}
\name{tibble}
\alias{tibble}
\title{Tibble}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 3 columns.
}
\usage{
data(tibble)
}
\value{
\code{tbl_df}.
}
\description{
Example \code{tbl_df} object (a tibble).
}
\note{
Updated 2021-10-14.
}
\examples{
if (requireNamespace("tibble", quietly = TRUE)) {
    data(tibble)
    class(tibble)
    tibble::is_tibble(tibble)
    tibble
}
}
\keyword{datasets}
