% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{txse}
\alias{txse}
\title{Example transcript-level SummarizedExperiment}
\format{An object of class \code{SummarizedExperiment} with 6 rows and 4 columns.}
\usage{
txse
}
\description{
Transcript-level RNA-seq data set.
}
\examples{
data(txse)
txse
}
\keyword{datasets}
