% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{AcidTest-package}
\alias{AcidTest}
\alias{AcidTest-package}
\title{AcidTest}
\description{
Reference data for Acid Genomics unit tests.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r.acidgenomics.com/packages/acidtest/}
  \item \url{https://github.com/acidgenomics/r-acidtest/}
  \item Report bugs at \url{https://github.com/acidgenomics/r-acidtest/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
