% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\docType{data}
\name{band_instruments}
\alias{band_instruments}
\title{Band membership}
\format{An object of class \code{DataFrame} with 3 rows and 2 columns.}
\usage{
band_instruments
}
\description{
These data sets describe band members of the Beatles and Rolling Stones. They
are toy data sets that can be displayed in their entirety on a slide (e.g. to
demonstrate a join).
}
\examples{
data(band_instruments)
band_instruments
}
\seealso{
\code{help(topic = "band_instruments", package = "dplyr")}
}
\keyword{internal}
