% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrame.R
\docType{data}
\name{DataFrame}
\alias{DataFrame}
\title{Data frame}
\format{An object of class \code{DataFrame} with 4 rows and 4 columns.}
\usage{
DataFrame
}
\description{
Example Bioconductor S4 \code{DataFrame} object.
}
\note{
Updated 2019-07-30.
}
\examples{
if (requireNamespace("S4Vectors", quietly = TRUE)) {
    data(DataFrame)
    class(DataFrame)
    DataFrame
}
}
\keyword{datasets}
