% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFrame.R
\docType{data}
\name{DFrame}
\alias{DFrame}
\title{Data frame}
\format{An object of class \code{DFrame} with 4 rows and 4 columns.}
\usage{
DFrame
}
\description{
Example Bioconductor S4 \code{DFrame} object.
}
\note{
Updated 2020-01-18.
}
\examples{
if (requireNamespace("S4Vectors", quietly = TRUE)) {
    data(DFrame)
    class(DFrame)
    DFrame
}
}
\keyword{datasets}
