% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gr}
\alias{gr}
\title{Genomic ranges}
\format{An object of class \code{GRanges} of length 5.}
\usage{
gr
}
\description{
Example \code{GRanges} object.
}
\examples{
suppressPackageStartupMessages(library(GenomicRanges))
data(gr)
class(gr)
gr
}
\keyword{datasets}
