% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges.R
\docType{data}
\name{GRanges}
\alias{GRanges}
\title{Genomic ranges}
\format{An object of class \code{GRanges} of length 5.}
\usage{
GRanges
}
\description{
Example \code{GRanges} object.
}
\examples{
if (requireNamespace("GenomicRanges", quietly = TRUE)) {
    data(GRanges)
    class(GRanges)
    GRanges
}
}
\keyword{datasets}
