% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ir}
\alias{ir}
\title{Example IRanges
Integer ranges.}
\format{An object of class \code{IRanges} of length 3.}
\usage{
ir
}
\description{
Example IRanges
Integer ranges.
}
\examples{
data(ir)
ir
}
\keyword{datasets}
