% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRanges.R
\docType{data}
\name{IRanges}
\alias{IRanges}
\title{Integer ranges}
\format{
An object of class \code{IRanges} of length 3.
}
\usage{
data(IRanges)
}
\value{
\code{IRanges}.
}
\description{
Examples \code{IRanges} object.
}
\note{
Updated 2020-01-28.
}
\examples{
if (requireNamespace("IRanges", quietly = TRUE)) {
    data(IRanges)
    class(IRanges)
    IRanges
}
}
\keyword{datasets}
