% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRanges.R
\docType{data}
\name{IRanges}
\alias{IRanges}
\title{Integer ranges}
\format{An object of class \code{IRanges} of length 3.}
\usage{
IRanges
}
\description{
Examples \code{IRanges} object.
}
\note{
Updated 2019-07-30.
}
\examples{
if (requireNamespace("IRanges", quietly = TRUE)) {
    data(IRanges)
    class(IRanges)
    IRanges
}
}
\keyword{datasets}
