% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lfc}
\alias{lfc}
\title{Example log2 fold change matrix}
\format{An object of class \code{matrix} with 8 rows and 4 columns.}
\usage{
lfc
}
\description{
Example log2 fold change matrix
}
\examples{
data(lfc)
lfc
}
\keyword{datasets}
