% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\docType{data}
\name{matrix}
\alias{matrix}
\title{Matrix}
\format{An object of class \code{matrix} with 4 rows and 4 columns.}
\usage{
matrix
}
\description{
Example \code{matrix} object.
}
\note{
Updated 2019-07-30.
}
\examples{
data(matrix)
class(matrix)
matrix
}
\keyword{datasets}
