% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce}
\alias{sce}
\title{Single cell experiment}
\format{An object of class \code{SingleCellExperiment} with 500 rows and 100 columns.}
\usage{
sce
}
\description{
Example \code{SingleCellExperiment} containing gene-level RNA-seq data.
}
\examples{
data(sce)
class(sce)
sce
}
\seealso{
\code{scater::sc_example_counts}.
}
\keyword{datasets}
