% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.R
\docType{data}
\name{Seurat}
\alias{Seurat}
\title{Seurat}
\format{An object of class \code{Seurat} with 230 rows and 80 columns.}
\usage{
Seurat
}
\description{
Example \code{Seurat} (v3) object.
}
\examples{
if (requireNamespace("Seurat", quietly = TRUE)) {
    data(Seurat)
    class(Seurat)
    Seurat
}
}
\keyword{datasets}
