% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratMarkersPerCluster.R
\docType{data}
\name{SeuratMarkersPerCluster}
\alias{SeuratMarkersPerCluster}
\title{Example \code{SeuratMarkersPerCluster} object}
\format{
An object of class \code{SeuratMarkersPerCluster} of dimension 106 x 79 x 37 x 8 x 8 x 8.
}
\usage{
data(SeuratMarkersPerCluster)
}
\value{
\code{SeuratMarkersPerCluster}.
}
\description{
Example \code{SeuratMarkersPerCluster} object
}
\note{
Updated 2022-06-09.
}
\examples{
if (requireNamespace("pointillism", quietly = TRUE)) {
    data(SeuratMarkersPerCluster)
    print(SeuratMarkersPerCluster)
}
}
\keyword{datasets}
