% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellExperiment_Seurat.R
\docType{data}
\name{SingleCellExperiment_Seurat}
\alias{SingleCellExperiment_Seurat}
\title{Single cell experiment}
\format{An object of class \code{SingleCellExperiment} with 230 rows and 80 columns.}
\usage{
SingleCellExperiment_Seurat
}
\description{
Example \code{SingleCellExperiment} containing gene-level RNA-seq data.
}
\note{
Updated 2020-01-18.
}
\examples{
if (requireNamespace("SingleCellExperiment", quietly = TRUE)) {
    data(SingleCellExperiment_Seurat)
    class(SingleCellExperiment_Seurat)
    SingleCellExperiment_Seurat
}
}
\seealso{
\itemize{
\item \code{data(pbmc_small, package = "Seurat")}.
}
}
\keyword{datasets}
