% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellExperiment_splatter.R
\docType{data}
\name{SingleCellExperiment_splatter}
\alias{SingleCellExperiment_splatter}
\title{Single cell experiment}
\format{
An object of class \code{SingleCellExperiment} with 100 rows and 400 columns.
}
\usage{
data(SingleCellExperiment_splatter)
}
\value{
\code{SingleCellExperiment}.
}
\description{
Example \code{SingleCellExperiment} containing gene-level RNA-seq data simulated
by the splatter package.
}
\note{
Updated 2022-03-07.
}
\examples{
if (requireNamespace("SingleCellExperiment", quietly = TRUE)) {
    data(SingleCellExperiment_splatter)
    class(SingleCellExperiment_splatter)
    print(SingleCellExperiment_splatter)
}
}
\seealso{
\itemize{
\item \code{splatter::splatSimulate()}.
\item \code{data(sc_example_counts, package = "scater")}.
}
}
\keyword{datasets}
