% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseMatrix.R
\docType{data}
\name{sparseMatrix}
\alias{sparseMatrix}
\title{Sparse matrix}
\format{
An object of class \code{dgCMatrix} with 8 rows and 10 columns.
}
\usage{
data(sparseMatrix)
}
\value{
\code{sparseMatrix}.
}
\description{
Example \code{sparseMatrix} (\code{Matrix}, \code{dgCMatrix}) object.
}
\note{
Updated 2029-01-28.
}
\examples{
if (requireNamespace("Matrix", quietly = TRUE)) {
    data(sparseMatrix)
    class(sparseMatrix)
    print(sparseMatrix)
}
}
\keyword{datasets}
