% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperiment_transcripts.R
\docType{data}
\name{SummarizedExperiment_transcripts}
\alias{SummarizedExperiment_transcripts}
\title{Summarized experiment}
\format{An object of class \code{SummarizedExperiment} with 6 rows and 4 columns.}
\usage{
data(SummarizedExperiment_transcripts)
}
\value{
\code{SummarizedExperiment}.
}
\description{
Example \code{SummarizedExperiment} containing transcript-level RNA-seq data.
}
\note{
Updated 2020-01-28.
}
\examples{
if (requireNamespace("SummarizedExperiment", quietly = TRUE)) {
    data(SummarizedExperiment_transcripts)
    class(SummarizedExperiment_transcripts)
    SummarizedExperiment_transcripts
}
}
\keyword{datasets}
