% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntactic.R
\docType{data}
\name{syntactic}
\alias{syntactic}
\title{Syntactic names list}
\format{An object of class \code{list} of length 6.}
\usage{
syntactic
}
\description{
Example \code{list} data used to test syntactic naming functions.
}
\examples{
data(syntactic)
class(syntactic)
names(syntactic)
}
\keyword{datasets}
