% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{txse}
\alias{txse}
\title{Transcript-level summarized experiment}
\format{An object of class \code{SummarizedExperiment} with 6 rows and 4 columns.}
\usage{
txse
}
\description{
Example \code{SummarizedExperiment} containing transcript-level RNA-seq data.
}
\examples{
if (requireNamespace("SummarizedExperiment", quietly = TRUE)) {
    data(txse)
    class(txse)
    txse
}
}
\keyword{datasets}
